Finally!! The First working Version!!
1 Chronicles 29
 11 Thine, O LORD, is the greatness, and the power, and the glory, and
the victory, and the majesty: for all that is in the heaven and in the
earth is thine; thine is the kingdom, O LORD, and thou art exalted as
head above all.
 12 Both riches and honour come of thee, and thou reignest over all; and
in thine hand is power and might; and in thine hand it is to make great,
and to give strength unto all.
 13 Now therefore, our God, we thank thee, and praise thy glorious name.
 (KJV)